chore(mocks): Moved mocks to seperate folder
