--- src/components/ui/Archive/mocks.ts
@@ -1,289 +1,0 @@
-import type { CardData, SimpleCategory } from '@/components/ui/Card/types';
-
-export const mockProjects = [
-  {
-    title: 'Vibz',
-    relationTo: 'projects',
-    details: {
-      type: 'Website',
-    },
-    updatedAt: '2025-03-10T11:33:28.064Z',
-    createdAt: '2025-03-10T11:33:28.064Z',
-    thumbnail: {
-      alt: 'test',
-      prefix: 'media',
-      filename: 'vibz-thumbnail.webp',
-      mimeType: 'image/webp',
-      filesize: 44600,
-      width: 1024,
-      height: 768,
-      focalX: 50,
-      focalY: 50,
-      createdAt: '2025-03-04T17:03:05.027Z',
-      updatedAt: '2025-03-04T17:03:05.027Z',
-      id: '67c73249759c5d5297da6891',
-      url: '/api/media/file/vibz-thumbnail.webp',
-      thumbnailURL: null,
-    },
-    slug: 'vibz',
-    id: '67c1bd0b9fb50c2e22c139f5',
-    categories: [
-      {
-        slug: 'digital',
-        title: 'Digital',
-        id: '67cecdbc00d7591f31a0091b',
-      },
-    ],
-  },
-  {
-    title: 'Urban Bites',
-    relationTo: 'projects',
-    details: {
-      type: 'Website',
-    },
-    thumbnail: {
-      alt: 'test',
-      prefix: 'media',
-      filename: 'vibz-thumbnail.webp',
-      mimeType: 'image/webp',
-      filesize: 44600,
-      width: 1024,
-      height: 768,
-      focalX: 50,
-      focalY: 50,
-      createdAt: '2025-03-04T17:03:05.027Z',
-      updatedAt: '2025-03-04T17:03:05.027Z',
-      id: '67c73249759c5d5297da6891',
-      url: '/api/media/file/vibz-thumbnail.webp',
-      thumbnailURL: null,
-    },
-    slug: 'urban-bites',
-    id: '67c1bgg0b9fb50c2e22c139f12',
-    updatedAt: '2025-03-10T11:33:28.064Z',
-    createdAt: '2025-03-10T11:33:28.064Z',
-    categories: [
-      {
-        title: 'Digital',
-        slug: 'digital',
-        id: '67cecdbc00d7591f31a0091b',
-      },
-      {
-        title: 'Branding',
-        slug: 'branding',
-        id: '352552352fgwegewgfe',
-      },
-    ],
-  },
-  {
-    title: 'Toyota',
-    relationTo: 'projects',
-    updatedAt: '2025-03-10T11:33:28.064Z',
-    createdAt: '2025-03-10T11:33:28.064Z',
-    thumbnail: {
-      alt: 'test',
-      prefix: 'media',
-      filename: 'vibz-thumbnail.webp',
-      mimeType: 'image/webp',
-      filesize: 44600,
-      width: 1024,
-      height: 768,
-      focalX: 50,
-      focalY: 50,
-      createdAt: '2025-03-04T17:03:05.027Z',
-      updatedAt: '2025-03-04T17:03:05.027Z',
-      id: '67c73249759c5d5297da6891',
-      url: '/api/media/file/vibz-thumbnail.webp',
-      thumbnailURL: null,
-    },
-    slug: 'post-3',
-    id: '67c1bgg0b9fb50c2e22c139f13',
-    details: {
-      type: 'Campaign',
-    },
-    categories: [
-      {
-        title: 'Print',
-        slug: 'print',
-        id: '67cece0800d7591f31a00957',
-      },
-    ],
-  },
-] satisfies CardData[];
-
-export const mockPosts = [
-  {
-    title: 'Post 1',
-    slug: 'post-1',
-    relationTo: 'posts',
-    id: '352552352fgwegewrwqfrfqwfqwfqfwgfe',
-    excerpt: 'This is an example excerpt for post one.',
-    updatedAt: '2025-03-10T11:33:28.064Z',
-    createdAt: '2025-03-10T11:33:28.064Z',
-    thumbnail: {
-      alt: 'test',
-      prefix: 'media',
-      filename: 'vibz-thumbnail.webp',
-      mimeType: 'image/webp',
-      filesize: 44600,
-      width: 1024,
-      height: 768,
-      focalX: 50,
-      focalY: 50,
-      createdAt: '2025-03-04T17:03:05.027Z',
-      updatedAt: '2025-03-04T17:03:05.027Z',
-      id: '67c73249759c5d5297da6891',
-      url: '/api/media/file/vibz-thumbnail.webp',
-      thumbnailURL: null,
-    },
-    categories: [
-      {
-        title: 'Digital',
-        slug: 'digital',
-        id: '67cecdbc00d7591f31a0091b',
-        description: 'Digital desc',
-      },
-      {
-        title: 'Graphic Design',
-        slug: 'graphic-design',
-        id: '352552352fgwegewrwqfrfqwfqwfqfwgfe',
-        description: 'Graphic-design desc',
-      },
-    ],
-  },
-  {
-    title: 'Post 2',
-    slug: 'post-2',
-    relationTo: 'posts',
-    id: '352552352fgwegewrwqfrfqwfqwfqfwgfewgggewe',
-    excerpt: 'This is an example excerpt for post two.',
-    updatedAt: '2025-03-10T11:33:28.064Z',
-    createdAt: '2025-03-10T11:33:28.064Z',
-    thumbnail: {
-      alt: 'test',
-      prefix: 'media',
-      filename: 'vibz-thumbnail.webp',
-      mimeType: 'image/webp',
-      filesize: 44600,
-      width: 1024,
-      height: 768,
-      focalX: 50,
-      focalY: 50,
-      createdAt: '2025-03-04T17:03:05.027Z',
-      updatedAt: '2025-03-04T17:03:05.027Z',
-      id: '67c73249759c5d5297da6891',
-      url: '/api/media/file/vibz-thumbnail.webp',
-      thumbnailURL: null,
-    },
-    categories: [
-      {
-        title: 'Brading',
-        slug: 'branding',
-        id: '352552352fgwegewgfe',
-        description: 'Branding desc',
-      },
-    ],
-  },
-  {
-    title: 'Post 3',
-    slug: 'post-3',
-    relationTo: 'posts',
-    id: '352552352fgwegewrwqfrfqwfqwfqfwgfewgggewew223441212g',
-    excerpt: 'This is an example excerpt for post three.',
-    updatedAt: '2025-03-10T11:33:28.064Z',
-    createdAt: '2025-03-10T11:33:28.064Z',
-    thumbnail: {
-      alt: 'test',
-      prefix: 'media',
-      filename: 'vibz-thumbnail.webp',
-      mimeType: 'image/webp',
-      filesize: 44600,
-      width: 1024,
-      height: 768,
-      focalX: 50,
-      focalY: 50,
-      createdAt: '2025-03-04T17:03:05.027Z',
-      updatedAt: '2025-03-04T17:03:05.027Z',
-      id: '67c73249759c5d5297da6891',
-      url: '/api/media/file/vibz-thumbnail.webp',
-      thumbnailURL: null,
-    },
-    categories: [
-      {
-        title: 'Graphic Design',
-        slug: 'graphic-design',
-        id: '352552352fgwegewrwqfrfqwfqwfqfwgfe',
-        description: 'Graphic-design desc',
-      },
-    ],
-  },
-  {
-    title: 'Post 4',
-    relationTo: 'posts',
-    slug: 'post-4',
-    updatedAt: '2025-03-10T11:33:28.064Z',
-    createdAt: '2025-03-10T11:33:28.064Z',
-    id: '352552352fgwegewrwqfrfqwfqwfqfwgfewgggewew2232e2e441212g',
-    excerpt: 'This is an example excerpt for post four.',
-    thumbnail: {
-      alt: 'test',
-      prefix: 'media',
-      filename: 'vibz-thumbnail.webp',
-      mimeType: 'image/webp',
-      filesize: 44600,
-      width: 1024,
-      height: 768,
-      focalX: 50,
-      focalY: 50,
-      createdAt: '2025-03-04T17:03:05.027Z',
-      updatedAt: '2025-03-04T17:03:05.027Z',
-      id: '67c73249759c5d5297da6891',
-      url: '/api/media/file/vibz-thumbnail.webp',
-      thumbnailURL: null,
-    },
-    categories: [
-      {
-        title: 'Web Design',
-        slug: 'web-design',
-        id: '352552352fgwegewrwqfrfqwfqwfqfwgfeergee',
-        description: 'Web Design desc',
-      },
-    ],
-  },
-] satisfies CardData[];
-
-export const mockCategories = [
-  {
-    description:
-      'Creating visually appealing and functional designs for physical materials, such as brochures, posters, flyers, business cards, and packaging. I ensure your printed materials leave a lasting impression.',
-    slug: 'print',
-    slugLock: true,
-    title: 'Print',
-    id: '67cece0800d7591f31a00957',
-  },
-  {
-    description:
-      'Crafting visually compelling websites that are easy to navigate and responsive across all devices. I focus on layout, typography, and interactive elements to create websites that look great and work even better.',
-    slug: 'digital',
-    slugLock: true,
-    title: 'Digital',
-    id: '67cecdbc00d7591f31a0091b',
-  },
-  {
-    title: 'Branding',
-    slug: 'branding',
-    id: '352552352fgwegewgfe',
-    description: 'Branding desc',
-  },
-  {
-    title: 'Graphic Design',
-    slug: 'graphic-design',
-    id: '352552352fgwegewrwqfrfqwfqwfqfwgfe',
-    description: 'Grpahic Design desc',
-  },
-  {
-    title: 'Web Design',
-    slug: 'web-design',
-    id: '352552352fgwegewrwqfrfqwfqwfqfwgfeergee',
-    description: 'Web Design desc',
-  },
-] satisfies SimpleCategory[];
