+++ src/mocks/data/mockPosts.ts
@@ -1,0 +1,108 @@
+import { CardData } from '@/components/ui/Card/types';
+
+export const mockPosts = [
+  {
+    title: 'Post 1',
+    slug: 'post-1',
+    relationTo: 'posts',
+    id: '352552352fgwegewrwqfrfqwfqwfqfwgfe',
+    excerpt: 'This is an example excerpt for post one.',
+    updatedAt: '2025-03-10T11:33:28.064Z',
+    createdAt: '2025-03-10T11:33:28.064Z',
+    thumbnail: {
+      alt: 'test',
+      prefix: 'media',
+      filename: 'vibz-thumbnail.webp',
+      mimeType: 'image/webp',
+      filesize: 44600,
+      width: 1024,
+      height: 768,
+      focalX: 50,
+      focalY: 50,
+      createdAt: '2025-03-04T17:03:05.027Z',
+      updatedAt: '2025-03-04T17:03:05.027Z',
+      id: '67c73249759c5d5297da6891',
+      url: '/api/media/file/vibz-thumbnail.webp',
+      thumbnailURL: null,
+    },
+    categories: ['67cecdbc00d7591f31a0091b', 'VkmADmVTWg234214'],
+  },
+  {
+    title: 'Post 2',
+    slug: 'post-2',
+    relationTo: 'posts',
+    id: '352552352fgwegewrwqfrfqwfqwfqfwgfewgggewe',
+    excerpt: 'This is an example excerpt for post two.',
+    updatedAt: '2025-03-10T11:33:28.064Z',
+    createdAt: '2025-03-10T11:33:28.064Z',
+    thumbnail: {
+      alt: 'test',
+      prefix: 'media',
+      filename: 'vibz-thumbnail.webp',
+      mimeType: 'image/webp',
+      filesize: 44600,
+      width: 1024,
+      height: 768,
+      focalX: 50,
+      focalY: 50,
+      createdAt: '2025-03-04T17:03:05.027Z',
+      updatedAt: '2025-03-04T17:03:05.027Z',
+      id: '67c73249759c5d5297da6891',
+      url: '/api/media/file/vibz-thumbnail.webp',
+      thumbnailURL: null,
+    },
+    categories: ['352552352fgwegewgfe'],
+  },
+  {
+    title: 'Post 3',
+    slug: 'post-3',
+    relationTo: 'posts',
+    id: '352552352fgwegewrwqfrfqwfqwfqfwgfewgggewew223441212g',
+    excerpt: 'This is an example excerpt for post three.',
+    updatedAt: '2025-03-10T11:33:28.064Z',
+    createdAt: '2025-03-10T11:33:28.064Z',
+    thumbnail: {
+      alt: 'test',
+      prefix: 'media',
+      filename: 'vibz-thumbnail.webp',
+      mimeType: 'image/webp',
+      filesize: 44600,
+      width: 1024,
+      height: 768,
+      focalX: 50,
+      focalY: 50,
+      createdAt: '2025-03-04T17:03:05.027Z',
+      updatedAt: '2025-03-04T17:03:05.027Z',
+      id: '67c73249759c5d5297da6891',
+      url: '/api/media/file/vibz-thumbnail.webp',
+      thumbnailURL: null,
+    },
+    categories: ['VkmADmVTWg234214'],
+  },
+  {
+    title: 'Post 4',
+    relationTo: 'posts',
+    slug: 'post-4',
+    updatedAt: '2025-03-10T11:33:28.064Z',
+    createdAt: '2025-03-10T11:33:28.064Z',
+    id: '352552352fgwegewrwqfrfqwfqwfqfwgfewgggewew2232e2e441212g',
+    excerpt: 'This is an example excerpt for post four.',
+    thumbnail: {
+      alt: 'test',
+      prefix: 'media',
+      filename: 'vibz-thumbnail.webp',
+      mimeType: 'image/webp',
+      filesize: 44600,
+      width: 1024,
+      height: 768,
+      focalX: 50,
+      focalY: 50,
+      createdAt: '2025-03-04T17:03:05.027Z',
+      updatedAt: '2025-03-04T17:03:05.027Z',
+      id: '67c73249759c5d5297da6891',
+      url: '/api/media/file/vibz-thumbnail.webp',
+      thumbnailURL: null,
+    },
+    categories: ['V181OQVabewdwd12', '67cece0800d7591f31a00957'],
+  },
+] satisfies CardData[];
