+++ src/mocks/data/mockProjects.ts
@@ -1,0 +1,93 @@
+import { CardData } from '@/components/ui/Card/types';
+
+export const mockProjects = [
+  {
+    title: 'Vibz',
+    relationTo: 'projects',
+    details: {
+      type: 'Website',
+    },
+    updatedAt: '2025-03-10T11:33:28.064Z',
+    createdAt: '2025-03-10T11:33:28.064Z',
+    thumbnail: {
+      alt: 'test',
+      prefix: 'media',
+      filename: 'vibz-thumbnail.webp',
+      mimeType: 'image/webp',
+      filesize: 44600,
+      width: 1024,
+      height: 768,
+      focalX: 50,
+      focalY: 50,
+      createdAt: '2025-03-04T17:03:05.027Z',
+      updatedAt: '2025-03-04T17:03:05.027Z',
+      id: '67c73249759c5d5297da6891',
+      url: '/api/media/file/vibz-thumbnail.webp',
+      thumbnailURL: null,
+    },
+    slug: 'vibz',
+    id: '67c1bd0b9fb50c2e22c139f5',
+    categories: ['67cecdbc00d7591f31a0091b', 'V181OQVabewdwd12'],
+  },
+  {
+    title: 'Urban Bites',
+    relationTo: 'projects',
+    details: {
+      type: 'Website',
+    },
+    thumbnail: {
+      alt: 'test',
+      prefix: 'media',
+      filename: 'vibz-thumbnail.webp',
+      mimeType: 'image/webp',
+      filesize: 44600,
+      width: 1024,
+      height: 768,
+      focalX: 50,
+      focalY: 50,
+      createdAt: '2025-03-04T17:03:05.027Z',
+      updatedAt: '2025-03-04T17:03:05.027Z',
+      id: '67c73249759c5d5297da6891',
+      url: '/api/media/file/vibz-thumbnail.webp',
+      thumbnailURL: null,
+    },
+    slug: 'urban-bites',
+    id: '67c1bgg0b9fb50c2e22c139f12',
+    updatedAt: '2025-03-10T11:33:28.064Z',
+    createdAt: '2025-03-10T11:33:28.064Z',
+    categories: ['67cecdbc00d7591f31a0091b', '352552352fgwegewgfe', 'V181OQVabewdwd12'],
+  },
+  {
+    title: 'Toyota',
+    relationTo: 'projects',
+    updatedAt: '2025-03-10T11:33:28.064Z',
+    createdAt: '2025-03-10T11:33:28.064Z',
+    thumbnail: {
+      alt: 'test',
+      prefix: 'media',
+      filename: 'vibz-thumbnail.webp',
+      mimeType: 'image/webp',
+      filesize: 44600,
+      width: 1024,
+      height: 768,
+      focalX: 50,
+      focalY: 50,
+      createdAt: '2025-03-04T17:03:05.027Z',
+      updatedAt: '2025-03-04T17:03:05.027Z',
+      id: '67c73249759c5d5297da6891',
+      url: '/api/media/file/vibz-thumbnail.webp',
+      thumbnailURL: null,
+    },
+    slug: 'post-3',
+    id: '67c1bgg0b9fb50c2e22c139f13',
+    details: {
+      type: 'Campaign',
+    },
+    categories: [
+      '67cecdbc00d7591f31a0091b',
+      '352552352fgwegewgfe',
+      'VkmADmVTWg234214',
+      '67cece0800d7591f31a00957',
+    ],
+  },
+] satisfies CardData[];
