+++ src/payload/collections/Services/index.ts
@@ -286,16 +286,5 @@ export const Services: CollectionConfig = {
         position: 'sidebar',
       },
     },
-    {
-      name: 'categories',
-      type: 'relationship',
-      relationTo: 'categories',
-      label: 'Categories',
-      hasMany: true,
-      required: true,
-      admin: {
-        position: 'sidebar',
-      },
-    },
   ],
 };
